requirements.txt
streamlit
ollama
chromadb
python-dotenv
requests
sseclient-py
langchain-ollama
langchain-core
langchain-community
langchain
pydantic
